removed SmartTrades V1 added some enums and changed optional parameters to nullable
