+++ XCommas.Net/XCommas.Net/Objects/SmartTrades/EnabledValue.cs
@@ -1,0 +1,12 @@
+﻿using Newtonsoft.Json;
+
+namespace XCommas.Net.Objects.SmartTrades
+{
+    public class EnabledValue
+    {
+        [JsonProperty("enabled")]
+        public bool Enabled { get; set; }
+        [JsonProperty("value")]
+        public double? Value { get; set; }
+    }
+}
