+++ XCommas.Net/XCommas.Net/Objects/SmartTrades/Leverage.cs
@@ -1,0 +1,11 @@
+﻿using Newtonsoft.Json;
+
+namespace XCommas.Net.Objects.SmartTrades
+{
+    public class Leverage
+    {
+        //    "type": "waiting_targets",
+        [JsonProperty("enabled")]
+        public bool Enabled { get; set; }
+    }
+}
