+++ XCommas.Net/XCommas.Net/Objects/SmartTrades/OrderDirection.cs
@@ -1,0 +1,12 @@
+﻿using System.Runtime.Serialization;
+
+namespace XCommas.Net.Objects.SmartTrades
+{
+    public enum OrderDirection
+    {
+        [EnumMember(Value = "asc")]
+        Asc,
+        [EnumMember(Value = "desc")]
+        Desc
+    }
+}
