+++ XCommas.Net/XCommas.Net/Objects/SmartTrades/OrderType.cs
@@ -1,0 +1,12 @@
+﻿using System.Runtime.Serialization;
+
+namespace XCommas.Net.Objects.SmartTrades
+{
+    public enum OrderType
+    {
+        [EnumMember(Value = "market")]
+        Market,
+        [EnumMember(Value = "limit")]
+        Limit
+    }
+}
