+++ XCommas.Net/XCommas.Net/Objects/SmartTrades/PriceType.cs
@@ -1,0 +1,14 @@
+﻿using System.Runtime.Serialization;
+
+namespace XCommas.Net.Objects.SmartTrades
+{
+    public enum PriceType
+    {
+        [EnumMember(Value = "bid")]
+        Bid,
+        [EnumMember(Value = "ask")]
+        Ask,
+        [EnumMember(Value = "last")]
+        Last
+    }
+}
