+++ XCommas.Net/XCommas.Net/Objects/SmartTrades/StopLoss.cs
@@ -1,0 +1,24 @@
+﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace XCommas.Net.Objects.SmartTrades
+{
+    public class StopLoss
+    {
+        [JsonProperty("enabled")]
+        public bool Enabled { get; set; }
+        [JsonConverter(typeof(StringEnumConverter))]
+        [JsonProperty("order_type")]
+        public OrderType OrderType { get; set; }
+        [JsonProperty("editable")]
+        public bool Editable { get; set; }
+        [JsonProperty("price")]
+        public Amount Price { get; set; }
+        [JsonProperty("conditional")]
+        public Conditional Conditional { get; set; }
+        [JsonProperty("timeout")]
+        public EnabledValue Timeout { get; set; }
+        [JsonProperty("status")]
+        public Status Status { get; set; }
+    }
+}
