+++ XCommas.Net/XCommas.Net/Objects/SmartTrades/TradeType.cs
@@ -1,0 +1,12 @@
+﻿using System.Runtime.Serialization;
+
+namespace XCommas.Net.Objects.SmartTrades
+{
+    public enum TradeType
+    {
+        [EnumMember(Value = "buy")]
+        Buy,
+        [EnumMember(Value = "sell")]
+        Sell
+    }
+}
